move util functions from muni types to utils folder
